fix flow and refactoring
