+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".Screens.MainActivity"
+            android:name=".Screens.SplashActivity"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -25,6 +25,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".Screens.MainActivity"
+            android:screenOrientation="portrait" />
         <activity
             android:name=".Screens.ShowARActivity"
             android:screenOrientation="portrait" />
