+++ app/src/main/res/layout/activity_main.xml
@@ -30,7 +30,7 @@
             android:layout_alignParentTop="true"
             android:layout_centerHorizontal="true"
             android:layout_marginTop="140dp"
-            android:text="AR" />
+            android:text="@string/ar" />
 
     </RelativeLayout>
