+++ app/src/main/res/layout/activity_show_ar.xml
@@ -11,7 +11,7 @@
         android:layout_alignParentEnd="true"
         android:layout_marginBottom="39dp"
         android:layout_marginEnd="25dp"
-        android:text="表示" />
+        android:text="@string/show" />
 
 
 </RelativeLayout>
