+++ app/src/main/res/layout/activity_splash.xml
@@ -9,21 +9,19 @@
         android:layout_height="wrap_content"
         android:background="@color/colorPrimary">
 
-
     </android.support.v7.widget.Toolbar>
 
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <Button
-            android:id="@+id/finish"
+        <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/exit"
-            android:layout_alignParentBottom="true"
+            android:text="@string/app_name"
+            android:layout_alignParentTop="true"
             android:layout_centerHorizontal="true"
-            android:layout_marginBottom="68dp" />
+            android:layout_marginTop="201dp" />
 
     </RelativeLayout>
